requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
feedparser>=6.0.0
deep-translator>=1.11.0
nba_api>=1.4.0
scipy>=1.11.0
plotly>=5.18.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
